+++ src/main/helpers/ApplicationStateHelper.ts
@@ -3,6 +3,8 @@ import { GenericTseApplicationType } from '../definitions/complexTypes/genericTs
 import { LinkStatus } from '../definitions/links';
 import CollectionUtils from '../utils/CollectionUtils';
 
+import { isResponseToTribunalRequired } from './GenericTseApplicationHelper';
+
 /**
  * Get application state for current user
  * @param app application
@@ -19,4 +21,21 @@ export const getApplicationState = (app: GenericTseApplicationType, user: UserDe
     return LinkStatus.IN_PROGRESS;
   }
   return LinkStatus.NOT_STARTED_YET;
+};
+
+/**
+ * Get new application state after viewed
+ * @param app application
+ * @param user current user
+ */
+export const getApplicationStatusAfterViewed = (app: GenericTseApplicationType, user: UserDetails): LinkStatus => {
+  const currentState = getApplicationState(app, user);
+  if (currentState === LinkStatus.NOT_VIEWED) {
+    return LinkStatus.VIEWED;
+  } else if (currentState === LinkStatus.UPDATED) {
+    return LinkStatus.IN_PROGRESS;
+  } else if (currentState === LinkStatus.NOT_STARTED_YET && isResponseToTribunalRequired(app, user)) {
+    return LinkStatus.IN_PROGRESS;
+  }
+  return undefined;
 };
